+++ R/zzz.r
@@ -1,21 +1,20 @@
 .onLoad <- function(libname, pkgname) {
-
-  # option: preferred font?
-  op <- options()
-  op.themes360info <- list(
-    # one of "itc", "libre" or "none"
-    themes360info.franklin.pref <- "itc"
-  )
-
-  # add any options that haven't been set by the user
-  toset <- !(names(op.themes360info) %in% names(op))
-  if (any(toset)) options(op.themes360info[toset])
-
-  # register fonts on package load (it can be called manually to re-register!)
-  font_choice <- getOption("themes360info.franklin.pref")
-  if (is.null(font_choice) || font_choice != "none") {
-    register_360fonts()
-  } else {
-    options("themes360info.franklin.loaded" = "none")
-  }
+  systemfonts::register_font(
+    name = "Headline 360info",
+    plain = system.file("extdata", "LibreFranklin-Black.ttf",
+      package = "themes360info"))
+  systemfonts::register_font(
+    name = "Subhead 360info",
+    plain = system.file("extdata", "LibreFranklin-Bold.ttf",
+      package = "themes360info"))
+  systemfonts::register_font(
+    name = "Body 360info",
+    plain = system.file("extdata", "LibreFranklin-Light.ttf",
+      package = "themes360info"), 
+    italic = system.file("extdata", "LibreFranklin-LightItalic.ttf",
+      package = "themes360info"),
+    bold = system.file("extdata", "LibreFranklin-Bold.ttf",
+      package = "themes360info"),
+    bolditalic = system.file("extdata", "LibreFranklin-BoldItalic.ttf",
+      package = "themes360info"))
 }